+++ src/languageservice/serviceclient.ts
@@ -26,8 +26,8 @@ import * as LanguageServiceContracts from '../models/contracts/languageService';
 import { IConfig } from '../languageservice/interfaces';
 import { exists } from '../utils/utils';
 import { env } from 'process';
-import { getAppDataPath, getAzureAuthLibraryConfig, getEnableSqlAuthenticationProviderConfig } from '../azure/utils';
-import { AuthLibrary } from '../models/contracts/azure';
+import { getAppDataPath, getEnableConnectionPoolingConfig, getEnableSqlAuthenticationProviderConfig } from '../azure/utils';
+import { serviceName } from '../azure/constants';
 
 const STS_OVERRIDE_ENV_VAR = 'MSSQL_SQLTOOLSSERVICE';
 
@@ -153,7 +153,7 @@ export default class SqlToolsServiceClient {
 
 	// gets or creates the singleton SQL Tools service client instance
 	public static get instance(): SqlToolsServiceClient {
-		if (this._instance === undefined) {
+		if (SqlToolsServiceClient._instance === undefined) {
 			let config = new ExtConfig();
 			let vscodeWrapper = new VscodeWrapper();
 			let logLevel: LogLevel = LogLevel[Utils.getConfigTracingLevel() as keyof typeof LogLevel];
@@ -167,9 +167,9 @@ export default class SqlToolsServiceClient {
 				decompressProvider);
 			let serviceProvider = new ServerProvider(downloadProvider, config, serverStatusView);
 			let statusView = new StatusView(vscodeWrapper);
-			this._instance = new SqlToolsServiceClient(config, serviceProvider, logger, statusView, vscodeWrapper);
+			SqlToolsServiceClient._instance = new SqlToolsServiceClient(config, serviceProvider, logger, statusView, vscodeWrapper);
 		}
-		return this._instance;
+		return SqlToolsServiceClient._instance;
 	}
 
 	// initialize the SQL Tools Service Client instance by launching
@@ -405,14 +405,19 @@ export default class SqlToolsServiceClient {
 			}
 
 			// Send application name and path to determine MSAL cache location
-			serverArgs.push('--application-name', 'code');
+			serverArgs.push('--application-name', serviceName);
 			serverArgs.push('--data-path', getAppDataPath());
 
 			// Enable SQL Auth Provider registration for Azure MFA Authentication
 			const enableSqlAuthenticationProvider = getEnableSqlAuthenticationProviderConfig();
-			const azureAuthLibrary = getAzureAuthLibraryConfig();
-			if (azureAuthLibrary === AuthLibrary.MSAL && enableSqlAuthenticationProvider) {
+			if (enableSqlAuthenticationProvider) {
 				serverArgs.push('--enable-sql-authentication-provider');
+			}
+
+			// Enable Connection pooling to improve connection performance
+			const enableConnectionPooling = getEnableConnectionPoolingConfig();
+			if (enableConnectionPooling) {
+				serverArgs.push('--enable-connection-pooling');
 			}
 
 			// Send Locale for sqltoolsservice localization