+++ src/objectExplorer/objectExplorerUtils.ts
@@ -78,7 +78,7 @@ export class ObjectExplorerUtils {
 		return LocalizedConstants.defaultDatabaseLabel;
 	}
 
-	public static isFirewallError(errorMessage: string): boolean {
-		return errorMessage.includes(Constants.firewallErrorMessage);
+	public static isFirewallError(errorCode: number): boolean {
+		return errorCode === Constants.errorFirewallRule;
 	}
 }